Skip apps/instance_reporting.go:45 because of Eirini for now
In Diego, it brings down the existing instance, then goes to
start the new instance and then fails with the expected
"Error starting instances: 'Insufficient resources: memory"
in a loop. Killing it leaves you with the expected 0
instances, and downtime.

In Eirini it fails earlier with "memory quota_exceeded", and
exit(1). This doesn't touch the app instances, and there's no
downtime, which is a better behaviour than Diego.